% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My first function}
\usage{
hello(x)
}
\arguments{
\item{x}{input for your name or anything}
}
\value{
The output from code
}
\description{
My first function
}
\examples{
\dontrun{hello("Chunhui")}

}
